Renames core/watcher/eventqueue to core/watcher/eventsource to reflect the name of the indirection on which these watcher foundations are based.
